Refactor parsing classses to uniform behaviours
