+++ lib/parsing/CsvFileParser.ts
@@ -3,34 +3,17 @@ import {OptionValues} from 'commander';
 import Line from './Line';
 import {logger} from '../tools/Logger';
 import readline from 'readline';
-import {createReadStream} from 'fs';
+import {createReadStream, PathLike} from 'fs';
 import chalk from 'chalk';
 
 export default class CsvFileParser extends AbstractParser {
 
-  separator: string;
-  from?: number;
-  to?: number;
-
   constructor(options: OptionValues) {
     super(options);
-    this.separator = options.separator ? options.separator : ';';
-    this.from = options.from;
-    this.to = options.to;
   }
 
-  parse(file: string): Promise<Line[]> {
-    return new Promise<Line[]>((resolve, reject) => {
-      logger.info('--------------------------------------------------------------------------------');
-      logger.info('Phase: "Parsing"');
-      logger.info(` - file: ${file}`);
-      logger.info(` - from: ${this.from}`);
-      logger.info(` - separator: ${this.separator}`);
-      logger.info(` - to: ${this.to}`);
-      logger.info();
-
-      const hrStart: [number, number] = process.hrtime();
-
+  extractLines(file: PathLike): Promise<Line[]> {
+    return new Promise<Line[]>((resolve) => {
       let index = 1;
       const lines: Line[] = [];
 
@@ -58,14 +41,7 @@ export default class CsvFileParser extends AbstractParser {
         index++;
       });
 
-      rl.on('close', () => {
-        logger.info();
-        const hrEnd: [number, number] = process.hrtime(hrStart);
-        logger.info(`Execution time (hr): ${hrEnd[0]}s ${hrEnd[1] / 1000000}ms`);
-        logger.info();
-
-        resolve(lines);
-      });
+      rl.on('close', () => resolve(lines));
     });
   }
 }
